This new version of selector adds a caching feature. This cache contains the pixel buffer of images to be displayed (faster), it is populated by image pixel buffer displayed and adds very little overhead compared to the former version. when images are re-displayed the cache is activated and the display is much faster( .1 sec compared to 5 seconds on a netbook).
By default the cache is activated at 100 MBytes representing arround 100 images of cache, on a netbook screen
it can be changed in the configuration file but no GUI exists at the moment for modifying it.
Don't be affrayed if you see selector grow at 120 MB of RAM ... above this can be a programming error.
enjoy this new version !!!

TODO: populate the cache with the new images freshly added to the repository as they will be displayed ...